Initialise nisol correctly if do.swap is TRUE
We use nisol as both an input vector (to tell us if do.swap is TRUE)
and an output vector.  After we have determined the correct value of
do.swap we then need to reinitialise nisol for output.  Only the first
entry of the array is used for input, and for output purposes nisol is
assumed to be zeroed, so just zero the first element.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/sprint@69 edb9625f-4e0d-4859-8d74-9fd3b1da38cb