Correctly initialise nisol on worker processes and gather result
The analysis code assumes that nisol has n_clusters entries, but we
previously only allocated space for 1 element.  This could result in
out of bound array accesses.  Fix this by allocating space for all
n_clusters on every process.  This is arguably not quite right, each
process only looks at n_clusters/nprocs entries so the array could be
smaller.  But that would require significantly more changes in the
logic.  This implementation relies on n_clusters not being too big.

As a result of this, we are now able to gather the results from nisol
back to the master process without causing out of bounds memory
accesses, so uncomment the code that does this.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/sprint@71 edb9625f-4e0d-4859-8d74-9fd3b1da38cb