Allocate function vector correctly on worker processes
To communicate the R function supplied by the user in papply, we
allocate a STRSXP that contains all the lines in the function and
broadcast those lines to all processes from the master.  Previously,
this object was allocated before the number of lines in the function
had been broadcast.  The size was thus dependent on whatever random
values happened to be on the stack.  Fix the problem by moving
allocation after the broadcast.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/sprint@74 edb9625f-4e0d-4859-8d74-9fd3b1da38cb